+++ crates/orchestrator/src/tests/common/mod.rs
@@ -5,7 +5,6 @@ use std::sync::Arc;
 
 use ::uuid::Uuid;
 use aws_config::meta::region::RegionProviderChain;
-use aws_sdk_sqs::types::Message;
 use constants::*;
 use da_client_interface::MockDaClient;
 use mongodb::Client;
@@ -92,7 +91,6 @@ pub async fn drop_database() -> color_eyre::Result<()> {
 }
 
 // SQS structs & functions
-// =============================================================
 
 pub async fn create_sqs_queues() -> color_eyre::Result<()> {
     let sqs_client = get_sqs_client().await;
@@ -100,10 +98,10 @@ pub async fn create_sqs_queues() -> color_eyre::Result<()> {
     // Dropping sqs queues
     let list_queues_output = sqs_client.list_queues().send().await?;
     let queue_urls = list_queues_output.queue_urls();
-    println!("Found {} queues", queue_urls.len());
+    log::debug!("Found {} queues", queue_urls.len());
     for queue_url in queue_urls {
         match sqs_client.delete_queue().queue_url(queue_url).send().await {
-            Ok(_) => println!("Successfully deleted queue: {}", queue_url),
+            Ok(_) => log::debug!("Successfully deleted queue: {}", queue_url),
             Err(e) => eprintln!("Error deleting queue {}: {:?}", queue_url, e),
         }
     }
@@ -114,39 +112,13 @@ pub async fn create_sqs_queues() -> color_eyre::Result<()> {
     Ok(())
 }
 
-pub async fn list_messages_in_queue(queue_name: String) -> color_eyre::Result<Vec<Message>> {
-    let sqs_client = get_sqs_client().await;
-    let mut all_messages = Vec::new();
-    let mut continue_receiving = true;
-
-    while continue_receiving {
-        let receive_message_output = sqs_client
-            .receive_message()
-            .queue_url(format!("http://sqs.ap-south-1.localhost.localstack.cloud:4566/000000000000/{}", queue_name))
-            .max_number_of_messages(10) // SQS allows receiving up to 10 messages at a time
-            .visibility_timeout(30) // Hide received messages for 30 seconds
-            .wait_time_seconds(20) // Enable long polling, wait up to 20 seconds for messages
-            .send()
-            .await?;
-
-        let messages = receive_message_output.messages();
-        if messages.is_empty() {
-            continue_receiving = false;
-        } else {
-            all_messages.extend(messages.iter().cloned());
-        }
-    }
-
-    Ok(all_messages)
-}
-
 async fn get_sqs_client() -> aws_sdk_sqs::Client {
     let region_provider = RegionProviderChain::default_provider().or_else("ap-south-1");
     let config = aws_config::from_env().region(region_provider).load().await;
     aws_sdk_sqs::Client::new(&config)
 }
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Debug)]
 pub struct MessagePayloadType {
     pub(crate) id: Uuid,
 }