+++ CHANGELOG.md
@@ -15,6 +15,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Changed
 
+- GitHub's coverage CI yml file for localstack and db testing.
+- Orchestrator :Moved TestConfigBuilder to `config.rs` in tests folder.
+
 ## Removed
 
 - `fetch_from_test` argument